% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighthouse_install.R
\name{lighthouse_install}
\alias{lighthouse_install}
\title{Lighthouse_install Function}
\usage{
lighthouse_install(prefix = "")
}
\arguments{
\item{prefix}{You can specify where you want Lighthouse to installed on your machine.
lighthouse_install()}
}
\description{
This function allows you to install Lighthouse on your local machine.
}
\examples{
\dontrun{
lighthouse_install()
}
}
