% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighthouse.R
\name{lighthouse}
\alias{lighthouse}
\title{Lighthouse Function}
\usage{
lighthouse(page, view = FALSE, keepFile = TRUE)
}
\arguments{
\item{page}{The Page you want to get analyzed by Lighthouse}

\item{view}{Logical. Do you want to open the HTML-File after the run? Default is FALSE.}

\item{keepFile}{Logical. Lighthouse generates a HTML-File with the results and saves it in your Workspace every Time you run it. If you don´t want this set keepFile to FALSE. Default is TRUE.
lighthouse()}
}
\description{
This function allows you to retrive Lighthouse Data.
Lighthouse will analyze the Page you specific in the request. Lighthouse will open a new Chrome Window and run various tests.
The output you will get consists of 5 numeric values: Progressive Web App, Performance, Accessibility, Best Practices, SEO. Each a number between 0 (very bad) and 100 (perfect).
Additional to this you will get a HTML-file in your Workspace containing all results of the Lighthouse test. So you can have a deeper look in various parts.
}
\examples{
\dontrun{
lighthouse("https://www.r-project.org/")
}
}
