% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighthouse_install.R
\name{lighthouse_install}
\alias{lighthouse_install}
\title{Lighthouse_install Function}
\usage{
lighthouse_install()
}
\description{
This function allows you to install Lighthouse on your lovcal machine.
lighthouse_install()
}
\examples{
{
lighthouse_install()
}
}
